Include osm-p2p-db-benchmark and display its results

+++ build-rnnodeapp.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 
 echo "Setting up...";
+rm -rf ./rnnodeapp;
 mkdir -p ./rnnodeapp;
 if [ "$1" = "benchmark" ]; then
   cp -r ./benchmark/rnnodeapp ./;
@@ -21,16 +22,20 @@ rm ./rnnodeapp/index.js;
 mv ./rnnodeapp/_index.js ./rnnodeapp/index.js;
 
 echo "Cleaning up...";
-declare -a keepThese=("hyperlog")
+declare -a keepThese=("hyperlog" "osm-p2p-db-benchmark")
 for x in "${keepThese[@]}"
 do
-  mv ./rnnodeapp/node_modules/$x ./rnnodeapp/$x;
+  if [ -e "./rnnodeapp/node_modules/$x" ]; then
+    mv ./rnnodeapp/node_modules/$x ./rnnodeapp/$x;
+  fi
 done
 rm -rf ./rnnodeapp/node_modules;
 mkdir -p ./rnnodeapp/node_modules;
 for x in "${keepThese[@]}"
 do
-  mv ./rnnodeapp/$x ./rnnodeapp/node_modules/$x;
+  if [ -e "./rnnodeapp/$x" ]; then
+    mv ./rnnodeapp/$x ./rnnodeapp/node_modules/$x;
+  fi
 done
 
 echo "Bundling into Android...";
